% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctis_get_micro_data.R
\name{ctis_get_micro}
\alias{ctis_get_micro}
\title{Download FB CTIS Daily Survey Data by date and file type.}
\usage{
ctis_get_micro(
  ctis_date,
  country = NA,
  ctis_type = "full",
  dest = "data/",
  ctis_username = Sys.getenv("CTIS_USERNAME"),
  ctis_password = Sys.getenv("CTIS_PASSWORD")
)
}
\arguments{
\item{ctis_date}{Date of CTIS Survey in `YYYY-MM-DD` format}

\item{country}{Country. Optional parameter. If supplied, only the records of
that particular country will be returned.}

\item{ctis_type}{CTIS file type. Can be either `full` or `partial`.}

\item{dest}{Destination Location.}

\item{ctis_username}{Username for the CTIS microdata API. Defaults to value
supplied in environment variable `CTIS_USERNAME`}

\item{ctis_password}{Password for the CTIS microdata API. Defaults to value
supplied in environment variable `CTIS_PASSWORD`}
}
\value{
A tibble.
}
\description{
Download FB CTIS Daily Survey Data by date and file type.
}
