% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctis_get_public_apidoc.R
\name{ctis_get_public_apidoc}
\alias{ctis_get_public_apidoc}
\title{Get CTIS Public API Documentation}
\usage{
ctis_get_public_apidoc()
}
\value{
A tibble with CTIS API Indicators Documentation.
}
\description{
Extracts API Documentation about CTIS Public API from from UMD's
[COVID-19 World Survey Open Data API](https://covidmap.umd.edu/api.html)
}
