% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctis_get_public_data.R
\name{ctis_get_public_data}
\alias{ctis_get_public_data}
\title{Get Aggregate Weighted and Unweighted CTIS Estimates.}
\usage{
ctis_get_public_data(
  indicator = "covid_vaccine",
  type = "daily",
  country = "India",
  region = NA,
  date_start = "2020-04-23",
  date_end = NA
)
}
\arguments{
\item{indicator}{CTIS Indicator. Examples - "covid", "mask", "vaccine_acpt", etc.,}

\item{type}{daily or smoothed}

\item{country}{Country. Examples - "Italy", "India", "Australia", etc.,}

\item{region}{Region. Optional.}

\item{date_start}{Start Date. Defaults to the beginning of survey 2020-04-23.}

\item{date_end}{End Date. Optional. If specified `all`, fetches till the
maximum available data i.e., 6 day lag.}
}
\value{
A tibble with indicator data for the country and date specified from
the Global COVID-19 Trends and Impact Survey
}
\description{
Provides Global COVID-19 Trends and Impact Survey's data for the
indicator, country, date and type specified from the  UMD's
[COVID-19 World Survey Open Data API](https://covidmap.umd.edu/api.html).
}
