% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctis_get_public_regions.R
\name{ctis_get_public_regions}
\alias{ctis_get_public_regions}
\title{Get CTIS Country Regions}
\usage{
ctis_get_public_regions()
}
\value{
A tibble with CTIS Country-Region pairs.
}
\description{
Provides CTIS Country Regions from UMD's
[COVID-19 World Survey Open Data API](https://covidmap.umd.edu/api.html)
}
\examples{
ctis_regions <- ctis_get_public_regions()
}
